+++ src/main/java/org/zalando/nakadi/repository/kafka/KafkaSettings.java
@@ -15,7 +15,7 @@ public class KafkaSettings {
     // https://github.com/apache/kafka/blob/d9206500bf2f99ce93f6ad64c7a89483100b3b5f/clients/src/main/java/org/apache
     // /kafka/clients/producer/ProducerConfig.java#L232
     private final int batchSize;
-    private final long lingerMs;
+    private final int lingerMs;
     private final boolean enableAutoCommit;
 
     private final int maxRequestSize;
@@ -23,7 +23,7 @@ public class KafkaSettings {
     @Autowired
     public KafkaSettings(@Value("${nakadi.kafka.request.timeout.ms}") final int requestTimeoutMs,
                          @Value("${nakadi.kafka.batch.size}") final int batchSize,
-                         @Value("${nakadi.kafka.linger.ms}") final long lingerMs,
+                         @Value("${nakadi.kafka.linger.ms}") final int lingerMs,
                          @Value("${nakadi.kafka.enable.auto.commit}") final boolean enableAutoCommit,
                          @Value("${nakadi.kafka.max.request.size}") final int maxRequestSize) {
         this.requestTimeoutMs = requestTimeoutMs;
@@ -41,7 +41,7 @@ public class KafkaSettings {
         return batchSize;
     }
 
-    public long getLingerMs() {
+    public int getLingerMs() {
         return lingerMs;
     }
